Handle disconnects with multiple pending requests.
If a server dies all the pending requests should switch over
to a different server.

+++ test/greeter/greeter.thrift
@@ -2,6 +2,6 @@ namespace rb.eventmachine Async
 
 service Greeter {
   string greeting(1:string name)
-  string delayed_greeting(1:string name, 2:i32 sleep_seconds)
+  string delayed_greeting(1:string name, 2:double sleep_seconds)
   oneway void yo(1:string name)
 }
